Load subcommand plug-ins at runtime
We use rubygems to load a subcommand plug-in gem with a file that can be
loaded at

    <command-path>/plugin.rb

The `command-path` is the namespace of the command converted to a path. For
example `Pod::Command` becomes `pod/command`.

The `plugin.rb` file should either define the subcommand(s) directly or
require the files for the subcommand(s). For example
[open_pod_bay](https://github.com/leshill/open_pod_bay) has the following
`plugin.rb`

    require 'pod/command/open'