[Command] Use only TTY check for ANSI output
CLAide includes by default the stub string mixin to prevent crashes (CLAide was
using the blue method but not checking for it). The additional initialization
time should be negligible and offset by the removal of conditional logic.

The specs are forced to run with ANSI disabled to prevent inconsistency issues.